automatically backed up by learn

+++ lib/bachelor.rb
@@ -80,7 +80,7 @@ def get_average_age_for_season(data, season)
   total_age = 0 
   total_contestants = 0
   data[season].each do |contestant_value|
-     total_age += contestant_value["age"].to_i
+     total_age += contestant_value["age"].to_f
      total_contestants += 1
   end 
   return (total_age / total_contestants)
